requirements.txt
PyYAML>=6.0
Jinja2>=3.1.0
click>=8.1.0
pydantic>=2.0.0
